Fix - opening storageResult with reindex flag

+++ src/main.cpp
@@ -2263,7 +2263,7 @@ bool DisconnectBlock(CBlock& block, CValidationState& state, CBlockIndex* pindex
         globalState->setRoot(uintToh256(pindex->pprev->hashStateRoot)); // lux
         globalState->setRootUTXO(uintToh256(pindex->pprev->hashUTXORoot)); // lux
 
-          if(!(!fLogEvents || !*pfClean)) {
+          if(fLogEvents && *pfClean) {
             pstorageresult->deleteResults(block.vtx);
             //pblocktree->EraseHeightIndex(pindex->nHeight);
         }
